challenge_backend/requirements.txt
appdirs==1.4.4
asgiref==3.5.1
astroid==2.9.1
autopep8==1.5.7
colorama==0.4.4
distlib==0.3.1
Django==4.0.4
django-cors-headers==3.12.0
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
filelock==3.0.12
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
platformdirs==2.4.1
pycodestyle==2.7.0
PyJWT==1.7.1
pylint==2.12.2
pytz==2022.1
six==1.15.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.0.1
tzdata==2022.1
virtualenv==20.4.4
wrapt==1.13.3
